mdrtb1.4.  cleaning register-type page completed.

+++ src/org/openmrs/module/mdrtb/MdrtbFactory.java
@@ -1597,9 +1597,7 @@ public class MdrtbFactory {
         for (PatientState lastState : lastStates){
         
             
-            if (lastState != null 
-                    && lastState.getState().getProgramWorkflow().getProgramWorkflowId() == programWorkflowState.getProgramWorkflow().getProgramWorkflowId()
-                    && !lastState.getVoided()) {
+            if (lastState != null  && lastState.getState().getProgramWorkflow().getProgramWorkflowId().intValue() == programWorkflowState.getProgramWorkflow().getProgramWorkflowId().intValue() && !lastState.getVoided()) {
                 
                     lastState.setEndDate(onDate);
                 //if nonsensical: void
